Create : #42 indexed 디렉토리 및 파일 분리
- mock data를 fixture로 분리
